app/requirements.txt
Flask==2.0.3
mysqlclient==2.1.1
python-dateutil==2.8.2